refactor: use balance() instead of value for skew
When computing the balance of the subtree rooted at a given node after
insertion, use balance() instead of the value.

Using the value avoids having to walk down both child pointers (which
may involve pointer-chasing into a memory load), but not using the value
allows non-copy value types to be stored in the tree prior to this
check without cloning.