feat: algebraic overlaps iterator
Adds an Iterator that yields all of the (interval, value) tuples in the
IntervalTree where the yielded interval overlaps with a given query
range.

This implementation always iterates over all nodes in the tree and uses
simple filtering to only return the appropriate tuples.

Also includes benchmarks iterating over a (deterministically
constructed) random tree for a given sample query range.